+++ bin/validator/src/db/sql/load_transaction.sql
@@ -1,6 +1,0 @@
-SELECT
-    submission_scheme,
-    submission_key_id,
-    sealed_transaction_inputs
-FROM validated_transactions
-WHERE id = ?1;